stig: tighten package check for RemoteAccessServices
The RemoteAccessServices check is intended to look for a set of
installed services / packages; however the pattern match rule did not
scope it's pattern to only match package names, so it would catch some
false positives.

This issue was masked by a prior bug in the pattern match that was fixed
in abb2237 ("changed hash script check into oval check and updated scap
schema") where it would never match anything because of the newline in
pattern -- that bug has been present since the initial publication of
the stig.

Signed-off-by: Steve Beattie <steve.beattie@chainguard.dev>
Ref: https://github.com/chainguard-dev/prodsec/issues/220