stig: reflow text
In prior updates, some of the textual descriptions were updated with
inconsistent formatting; re-flow the text descriptions to make them more
readable.

Note that these changes are mostly not user visible; the html report
generated by openscap reflows the text, the update here is mostly for
the convenience of people looking at the xml document directly. The
exception to this is the change to the xccdf_._rule_V_263659 "Script
Verfication" section; this text is defined to be pre-formatted html
and thus is emitted by openscap in its report as-is.

Signed-off-by: Steve Beattie <steve.beattie@chainguard.dev>