add ca-cert PR generator github workflow
Add a daily workflow that compares the sha256sum of the ca-certificates
bundle from the latest wolfi-base image with the sha256 hash embedded in
the stig profile and create a PR if the bundle has been updated.

This is based on the similar workflow that was prototyped in the
oscap-playground repo.

Signed-off-by: Steve Beattie <steve.beattie@chainguard.dev>