update-ca-cert workflow: fix version ref to create-pull-request action
The eter-evans/create-pull-request github action is correctly pinned to
a specific commit; however the version reference was copy-pasta'd
incorrectly, and so dependabot does not update it when proposing to
incorporate a newer version of the action.

See https://github.com/chainguard-dev/stigs/pull/35/files as an example
of dependabot not adjusting the version correctly.

fixes: b270309 ("add ca-cert PR generator github workflow")
Signed-off-by: Steve Beattie <steve.beattie@chainguard.dev>