RemoteAccessServices: remove unneeded state reference
oscap xccdf validate was failing on the RemoteAccessServices checks both
in the combined oval and in the individual definition, like so:

  OpenSCAP Error: Invalid OVAL Definition (5.11.1) content in RemoteAccessServicesTest.xml. [/home/build/src/source/oscap_source.c:380]
  <?xml version="1.0"?>
  oval:org.RemoteAccessServices:tst:1 - No state should be referenced when check_existence has a value of 'none_exist'.

When a check_existence test expects a value of 'none_exist' there should
not be a `state` referenced (or defined). Fix this by removing the
definition and reference to the state "oval:my_test:ste:1".

Signed-off-by: Steve Beattie <steve.beattie@chainguard.dev>